Add a hack to make the two-finger tap more reliable
When pressing with two fingers, we actually get, first,
a single touch, then the two-touches but with the coords
from the 2nd finger.

We need to save the location of the first touch, then
of the 2nd touch, and use that as a negative offset to
the broken 2nd touch events we'll get.